backend/app/requirements.txt
aiofiles==23.1.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cryptography==41.0.1
fastapi==0.99.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-aiplatform==1.27.1
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-resource-manager==1.10.2
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
vertexai==0.0.1
